Feature/improve parameter passing

See merge request julrich/QtPromise!30

+++ src/Deferred.cpp
@@ -4,7 +4,7 @@
 
 namespace QtPromise {
 
-QAtomicInteger<qint8> Deferred::m_metaTypesRegistered{0};
+QAtomicInteger<int> Deferred::m_metaTypesRegistered{0};
 
 Deferred::Deferred()
 	: QObject(nullptr), m_state(Pending), m_lock(QMutex::Recursive)
